Title: 사칙연산 함수, Time: 131ms, Memory: 24MB, Status: Passed - CodeTree

+++ 240731/사칙연산 함수/quadratic-operations-function.py
@@ -10,6 +10,6 @@ elif o == '-':
 elif o == '*':
     print(n,'=',a*c)
 elif o == '/':
-    print(n,'=',a/c)
+    print(n,'=',int(a/c))
 else:
     print('False')
